+++ app/controllers/user/likes_controller.rb
@@ -14,8 +14,10 @@ class User::LikesController < ApplicationController
 
   def destroy
     @review = Review.find(params[:review_id])
-    like = current_user.likes.find_by(review_id: @review.id)
-    like.destroy
+    if @review.liked_by?(current_user)
+      like = current_user.likes.find_by(review_id: @review.id)
+      like.destroy
+    end
     @review.user.point -= 1
     @review.user_rank_update(@review.user) # いいねを外したレビューを投稿したユーザーのランクをアップデート
   end
